requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.7.0,<3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
langchain>=0.1.0
langchain-core>=0.1.12,<0.2.0
langchain-openai==0.0.2
langsmith>=0.0.77,<0.1.0
openai>=1.3.0
motor==3.3.1
python-multipart==0.0.6
typing-extensions>=4.10,<5.0.0
anthropic==0.49.0
langgraph==0.0.15
langchain-community>=0.0.10
python-docx==0.8.11
# langchain-anthropic==0.3.9
# langgraph-checkpoint==2.0.19
# langgraph-prebuilt==0.1.2
# pydantic-settings==2.8.1 